fixup: address Copilot review feedback (PR #2152)
- Move subprocess tests from src/main.rs to tests/empty_stdin.rs
  using CARGO_BIN_EXE_delta for cross-platform binary discovery
- Set --pager=cat in paging_always test to prevent hang in CI
- Remove review artefact REVIEW-*.md from PR
