+++ src/main.rs
@@ -306,78 +306,3 @@ pub fn run_app(
     // `output_type` drop impl runs here
 }
 
-#[cfg(test)]
-mod main_tests {
-    use std::io::Write;
-    use std::process::{Command, Stdio};
-
-    /// Locate the delta binary in the cargo target directory.
-    /// During `cargo test`, the test binary is at target/debug/deps/git_delta-<hash>,
-    /// and the main binary is at target/debug/delta.
-    fn delta_binary() -> Option<std::path::PathBuf> {
-        let test_bin = std::env::current_exe().ok()?;
-        let mut path = test_bin;
-        path.pop(); // deps/
-        path.pop(); // debug/ or release/
-        path.push("delta");
-        if path.exists() { Some(path) } else { None }
-    }
-
-    #[test]
-    fn test_no_output_on_empty_stdin_when_stdout_piped() {
-        let delta_bin = delta_binary().expect(
-            "delta binary not found at target/debug/delta; run `cargo build --bin delta` first"
-        );
-
-        let mut child = Command::new(&delta_bin)
-            .arg("--no-gitconfig")
-            .stdin(Stdio::piped())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::null())
-            .spawn()
-            .expect("failed to spawn delta");
-
-        child.stdin.take().unwrap().write_all(b"").unwrap();
-        let output = child.wait_with_output().unwrap();
-
-        assert!(
-            output.status.success(),
-            "delta exited with error on empty stdin: {}",
-            String::from_utf8_lossy(&output.stderr)
-        );
-        assert_eq!(
-            output.stdout.len(),
-            0,
-            "Expected 0 bytes on empty stdin piped to delta, got {} bytes: {:?}",
-            output.stdout.len(),
-            String::from_utf8_lossy(&output.stdout)
-        );
-    }
-
-    #[test]
-    fn test_empty_stdin_does_not_hang_with_paging_always() {
-        let delta_bin = delta_binary().expect(
-            "delta binary not found at target/debug/delta; run `cargo build --bin delta` first"
-        );
-
-        let mut child = Command::new(&delta_bin)
-            .arg("--no-gitconfig")
-            .arg("--paging=always")
-            .stdin(Stdio::piped())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::null())
-            .spawn()
-            .expect("failed to spawn delta");
-
-        child.stdin.take().unwrap().write_all(b"").unwrap();
-        let output = child.wait_with_output().unwrap();
-
-        // --paging=always still starts the pager even when stdout is piped.
-        // The test verifies this doesn't hang and exits cleanly.
-        assert!(
-            output.status.success(),
-            "delta --paging=always exited with error on empty stdin: {}",
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
-}
